Library

Contents

Library

Contents

Functions

Index

Typedefs

Structs

General functions

Utils

Functions

Index

OpenGeoSysUncertaintyQuantification.OGS6ProjectParams

OpenGeoSysUncertaintyQuantification.StochasticOGS6Parameter

OpenGeoSysUncertaintyQuantification.dependend_parameter!

OpenGeoSysUncertaintyQuantification.dependend_tensor_parameter!

Typedefs

Structs

OpenGeoSysUncertaintyQuantification.OGS6ProjectParams — Type

```julia mutable struct OGS6ProjectParams```

Container for OpenGeoSys 6 Parameters

Fields

projectfile::String : Path to the OGS6 project file (path/to/project.prj)

simcall::String : Path to the OGS6 binary (path/to/ogs/bin/ogs)

additionalprojecfilespath::String : Path to the folder with additional project files (meshes & scripts) which gets copied to each realization folder

outputpath::String : Path to Result folder (path/to/stochprojectfolder/Res/)

postprocfiles::Vector{String} : Array of OGS6 postprocessing results either VTU or XDMF files

OpenGeoSysUncertaintyQuantification.StochasticOGS6Parameter — Type

mutable struct StochasticOGS6Parameter

Container for a stochastic parameter

Fields

path::String : OGS6 path definition (see Ogs6InputFileHandler)

valspec::Int : Value specifier (1 for scalar parameters, 1,...,nvals for tensor parameters)

dist::UnivariateDistribution : Univariate distribution (see )

lower_bound::Float64 : Lower bound for truncated distribution (all normal distributions must be truncated)

upper_bound::Float64 : Upper bound for truncated distribution (all normal distributions must be truncated)

General functions

Utils

OpenGeoSysUncertaintyQuantification.dependend_tensor_parameter! — Method

dependend_tensor_parameter!(modeldef::Ogs6ModelDef, stoparam::String, master_ind::Int, slave_ind::Int, depfunc)

Helper function to model a dependency of entries of a tensor parmeter such as the permeability tensor k.

Arguments

modeldef::Ogs6ModelDef: OSG6 Model Definition

stoparam::String: OGS Path to parameter

master_ind::Int: master index

slave_ind::Int: slave index

depfunc: dependency function

Example:

dependend_tensor_parameter!(
	modeldef, 
	path_to_permeability, 
	1,
	3, 
	x->0.5*x)

The above example sets the third entry of the permeability tensor to half of the value of the first entry.

OpenGeoSysUncertaintyQuantification.dependend_parameter! — Method

dependend_parameter!(modeldef::Ogs6ModelDef, masterstoparam::String, slavestoparam::String, master_ind::Int, slave_ind::Int, depfunc)

Helper function to model a dependency of two OGS6 parameter

Arguments

modeldef::Ogs6ModelDef: OSG6 Model Definition

masterstoparam::String: OGS Path to master parameter

slavestoparam::String: OGS Path to slave parameter

master_ind::Int: master index

slave_ind::Int: slave index

depfunc: dependency function

Example:

dependend_tensor_parameter!(
	modeldef, 
	path_to_permeability_id1, 
	path_to_permeability_id2, 
	1, 
	1, 
	x->x)

The above example couples two permeabilites of different material layers.