Library

Contents

Library

Contents

Index

Typedefs

Structs

General functions

Project setup

Utils

Utils adaptive sparse grid

Utils userfile

Index

OpenGeoSysUncertaintyQuantification.MonteCarloMorrisParams

OpenGeoSysUncertaintyQuantification.MonteCarloParams

OpenGeoSysUncertaintyQuantification.MonteCarloSobolParams

OpenGeoSysUncertaintyQuantification.OGS6ProjectParams

OpenGeoSysUncertaintyQuantification.OGSUQASG

OpenGeoSysUncertaintyQuantification.OGSUQMC

OpenGeoSysUncertaintyQuantification.OGSUQMCMorris

OpenGeoSysUncertaintyQuantification.OGSUQMCSobol

OpenGeoSysUncertaintyQuantification.OGSUQParams

OpenGeoSysUncertaintyQuantification.SampleMethodParams

OpenGeoSysUncertaintyQuantification.SparseGridParams

OpenGeoSysUncertaintyQuantification.StochasticOGS6Parameter

OpenGeoSysUncertaintyQuantification.StochasticOGSModelParams

DistributedSparseGrids.init

OpenGeoSysUncertaintyQuantification.dependend_parameter!

OpenGeoSysUncertaintyQuantification.dependend_tensor_parameter!

OpenGeoSysUncertaintyQuantification.generatePossibleStochasticParameters

OpenGeoSysUncertaintyQuantification.scalarwise_comparefct

OpenGeoSysUncertaintyQuantification.start!

OpenGeoSysUncertaintyQuantification.variance

OpenGeoSysUncertaintyQuantification.𝔼

Typedefs

OpenGeoSysUncertaintyQuantification.SampleMethodParams — Type

abstract type SampleMethodParams

Supertype for Sample method parameters.

Structs

OpenGeoSysUncertaintyQuantification.OGS6ProjectParams — Type

mutable struct OGS6ProjectParams

Container for OpenGeoSys 6 Parameters

Fields

projectfile::String : Path to the OGS6 project file (e.g. path/to/project.prj).

simcall::String : Path to the OGS6 binary (e.g. path/to/ogs/bin/ogs).

additionalprojecfilespath::String : Path to the folder with additional project files (meshes & scripts) which gets copied to each realization folder.

outputpath::String : Path to Result folder (e.g. path/to/stochprojectfolder/Res/).

postprocfiles::Vector{String} : Array of OGS6 postprocessing results containing either vtu files (readable by VTUFileHandler.VTUFile) or xdmf files (readable by XDMFFileHandler.XDMF3File).

OpenGeoSysUncertaintyQuantification.StochasticOGS6Parameter — Type

mutable struct StochasticOGS6Parameter

Container for the definition of a stochastic OGS6 parameter. For all distributions a lower and upper bound have to be provided. For Distributions.Uniform, this can be the interval $[a,b]$, for Distributions.Normal, proper bounds have to be provided.

Fields

path::String : OGS6 path definition (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath).

valspec::Int : Value specifier (1 for scalar parameters, $i$ for $i$-th value of a tensor parameters written as a vector in OGS6 e.g. [$a_1=a_{11}, a_2=a_{12}, a_3=a_{21}, a_4=a_{22}$]).

dist::UnivariateDistribution : Univariate distribution (see Distributions.UnivariateDistribution).

lower_bound::Float64 : Lower bound for truncated distribution (see Distributions.truncated).

upper_bound::Float64 : Upper bound for truncated distribution (see Distributions.truncated).

OpenGeoSysUncertaintyQuantification.StochasticOGSModelParams — Type

mutable struct StochasticOGSModelParams

Container defining the stochastic OGS6 model.

Fields

ogsparams::OGS6ProjectParams : OGS 6 project parameters.

stochparams::Vector{StochasticOGS6Parameter} : Vector defining the stochastic state space.

samplemethod::Type : Either DistributedSparseGrids.AdaptiveHierarchicalSparseGrid, DistributedMonteCarlo.MonteCarlo, DistributedMonteCarlo.MonteCarloSobol, or DistributedMonteCarlo.MonteCarloMorris.

num_local_workers::Int : Number of local workers to be added by Distributed.addprocs.

userfunctionfile::String : path to userfunction file.

file::String : path to file to write the StochasticOGSModelParams as XML-file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.SparseGridParams — Type

mutable struct SparseGridParams <: SampleMethodParams

Container for sparse grid paramters. For more information on the adaptive sparse grid, see the documentation.

Fields

N::Int : Dimension of the sparse grid, e.g. the stochastic state space.

CT::Type : Type of the collocation points, normally Float64.

RT::Type : Return type of the function fun calling the OGS6 binary. Can be customized to Float64, Vector{Float64}, Matrix{Float64}, XDMFFileHandler.XDMF3File, VTUFileHandler.VTUFile, or a custom data type which have to implement the AltInplaceOpsInterface.

pointprobs::Vector{Int} : Point set properties of the sparse grid, see DistributedSparseGrids.init and the point set example.

init_lvl::Int : Number of initial hierarchical levels before adaptive refinement starts.

maxlvl::Int : Number of additional hierarchical level after init_lvl.

tol::Float64 : tolerance for adaptive refinement.

file::String : path to file to write the StochasticOGSModelParams as XML-file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.MonteCarloParams — Type

mutable struct MonteCarloParams <: SampleMethodParams

Container for Monte Carlo parameters.

Fields

N::Int : Dimension of the sparse grid, e.g. the stochastic state space.

CT::Type : Type of the coordinates in the stochastic state space, normally Float64.

RT::Type : Return type of the function fun calling the OGS6 binary. Can be customized to Float64, Vector{Float64}, Matrix{Float64}, XDMFFileHandler.XDMF3File, VTUFileHandler.VTUFile, or a custom data type which have to implement the AltInplaceOpsInterface.

nshots::Int : Number of Monte Carlo snapshots.

tol::Float64 : tolerance (not used at the moment).

file::String : path to file to write the StochasticOGSModelParams as XML-file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.MonteCarloSobolParams — Type

mutable struct MonteCarloSobolParams <: SampleMethodParams

Container for Monte Carlo Sobol parameters.

Fields

N::Int : Dimension of the sparse grid, e.g. the stochastic state space.

CT::Type : Type of the coordinates in the stochastic state space, normally Float64.

RT::Type : Return type of the function fun calling the OGS6 binary. Can be customized to Float64, Vector{Float64}, Matrix{Float64}, XDMFFileHandler.XDMF3File, VTUFileHandler.VTUFile, or a custom data type which have to implement the AltInplaceOpsInterface.

nshots::Int : Number of Monte Carlo snapshots.

tol::Float64 : tolerance (not used at the moment).

file::String : path to file to write the StochasticOGSModelParams as XML-file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.MonteCarloMorrisParams — Type

mutable struct MonteCarloMorrisParams

Container for Monte Carlo Morris parameters.

Fields

N::Int : Dimension of the sparse grid, e.g. the stochastic state space.

CT::Type : Type of the coordinates in the stochastic state space, normally Float64.

RT::Type : Return type of the function fun calling the OGS6 binary. Can be customized to Float64, Vector{Float64}, Matrix{Float64}, XDMFFileHandler.XDMF3File, VTUFileHandler.VTUFile, or a custom data type which have to implement the AltInplaceOpsInterface.

ntrajectories::Int : Number of trajectories. Each trajectories triggers $N+1$ deterministic OGS6 runs.

lhs_sampling::Bool : If set to true a optimized Latin Hypercube design is used instead of Monte Carlo integration. Functionality provided by the LatinHypercubeSampling.jl package.

tol::Float64 : tolerance (not used at the moment).

file::String : path to file to write the StochasticOGSModelParams as XML-file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.OGSUQParams — Type

mutable struct OGSUQParams

Container for stochastic model parameters and sample method parameters.

Fields

stochasticmodelparams::StochasticOGSModelParams : stochastic model parameters.

samplemethodparams::SampleMethodParams : sample method parameters:

Can be instantiated by

OGSUQParams(
	stochasticmodelparams::StochasticOGSModelParams, 
	samplemethodparams::SampleMethodParams
	)

or

OGSUQParams(
	file_stochasticmodelparams::String, 
	file_samplemethodparams::String
	)

where file_stochasticmodelparams and file_samplemethodparams are pathes to xml files of StochasticOGSModelParams and SampleMethodParams, respectively, written by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.OGSUQASG — Type

mutable struct OGSUQASG

A stochastic OGS6 adaptive sparse grid model. The sparse grid is implemented by DistributedSparseGrids.jl. For more information, see the documentation. Instatiated by init function.

Fields

ogsuqparams::OGSUQParams : stochastic model and sample method parameters.

asg::DistributedSparseGrids.AdaptiveHierarchicalSparseGrid : instance of the adaptive sparse grid

OpenGeoSysUncertaintyQuantification.OGSUQMC — Type

mutable struct OGSUQMC

A stochastic OGS6 Monte Carlo model. The Monte Carlo Method is implemented by DistributedMonteCarlo.jl. Instatiated by init function.

Fields

ogsuqparams::OGSUQParams : stochastic model and sample method parameters.

asg::DistributedMonteCarlo.MonteCarlo : instance of the Monte Carlo integrator.

OpenGeoSysUncertaintyQuantification.OGSUQMCSobol — Type

mutable struct OGSUQMCSobol

A stochastic OGS6 Sobol model. Integration of the Sobol indices implemented by DistributedMonteCarlo.jl. Instatiated by init function.

Fields

ogsuqparams::OGSUQParams : stochastic model and sample method parameters.

mc::DistributedMonteCarlo.MonteCarloSobol : instance of the Monte Carlo Sobol integrator.

OpenGeoSysUncertaintyQuantification.OGSUQMCMorris — Type

mutable struct OGSUQMCMorris

A Stochastic OGS6 Morris model. Integration of the Sobol indices implemented by DistributedMonteCarlo.jl. Instatiated by init function.

Fields

ogsuqparams::OGSUQParams : stochastic model and sample method parameters.

mc::DistributedMonteCarlo.MonteCarloMorris : instance of the Monte Carlo Sobol integrator.

General functions

DistributedSparseGrids.init — Method

init(ogsuqparams::OGSUQParams)

Helper function to instantiate a stochastic OGS6 model. Return an object of type OGSUQASG, OGSUQMC, OGSUQMCSobol, or OGSUQMCMorris.

Arguments

ogsuqparams::OGSUQParams: Stochastic model parameters.

OpenGeoSysUncertaintyQuantification.start! — Function

start!(ogsuqasg::OGSUQASG, refinetohyperedges=false)

Starts the adaptive sparse grid sampling procedure.

Arguments

ogsuqasg::OGSUQASG: Stochastic OGS adaptive sparse grid model.

refinetohypercube::Bool: If set to true the sparse grid is refined such that all vertices of the n-cube spawn.

start!(ogsuqmc::OGSUQMC)

Starts the Monte Carlo sampling. Since Monte Carlo results are loaded each time expected value or variance is computed, only the result folder is checked for existing Monte Carlo snaphots. If there are any, the snapshot coordinates are reloaded.

Arguments

ogsuqmc::OGSUQMC: Stochastic OGS Monte Carlo model.

start!(ogsuqmc::OGSUQMCSobol)

Starts the Monte Carlo Sobol sampling.

Arguments

ogsuqmc::OGSUQMCSobol: Stochastic OGS Monte Carlo Sobol model.

start!(ogsuqmc::OGSUQMCMorris)

Starts the Monte Carlo Sobol sampling.

Arguments

ogsuqmc::OGSUQMCMorris: Stochastic OGS Monte Carlo Morris model.

OpenGeoSysUncertaintyQuantification.𝔼 — Function

𝔼(ogsuqmc::OGSUQMC)

Computes the expected value. Calls DistributedMonteCarlo.distributed_𝔼. All values are computet in-place. Only few values are stored in memory simultaneously so that the expected value of large sample sizes can be computed.

Arguments

ogsuqmc::OGSUQMC: Stochastic OGS Monte Carlo model.

𝔼(ogsuqasg::OGSUQASG)

Computes the expected value. Hereby, the physical model ogsuqasg.asg is evaluated against the multivariate pdf. Results are stored in memory multiple times. Feasible if all samples only occupy a fraction of the system memory.

Arguments

ogsuqasg::OGSUQASG: Stochastic OGS adaptive sparse grid model.

OpenGeoSysUncertaintyQuantification.variance — Function

variance(ogsuqasg::OGSUQASG)

Computes the expected value. Hereby, the physical model ogsuqasg.asg is evaluated against the multivariate pdf. Results are stored in memory multiple times. Feasible if all samples only occupy a fraction of the system memory.

Arguments

ogsuqasg::OGSUQASG: Stochastic OGS adaptive sparse grid model.

variance(ogsuqmc::OGSUQMC)

Computes the variance. Calls DistributedMonteCarlo.distributed_𝔼. All values are computet in-place. Only few values are stored in memory simultaneously so that the expected value of large sample sizes can be computed.

Arguments

ogsuqmc::OGSUQMC: Stochastic OGS Monte Carlo model.

Project setup

OpenGeoSysUncertaintyQuantification.generatePossibleStochasticParameters — Function

generatePossibleStochasticParameters(
	projectfile::String,
	file::String="./PossibleStochasticParameters.xml",
	keywords::Vector{String}=ogs_numeric_keyvals
	)

Helper function for initial setup of a stochastic project. Scans an existing OGS6 projectfile (.prj) for keywords indicating numeric values usable as stochastic input parameter. Returns a Vector of Strings with OGS6 pathes. Uses OGS6 Pathes for parameter indentification (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath).

Arguments

projectfile::String: Path to OGS6 project file (e.g. ./path/to/project.prj)

file::String: Path of output XML file which gets written by this function.

keywords::Vector{String}: List of keywords.

Utils

Utils adaptive sparse grid

OpenGeoSysUncertaintyQuantification.scalarwise_comparefct — Method

scalarwise_comparefct(rt,tolrt,mintol)

Criterion for refinement of the DistributedSparseGrids.AdaptiveHierarchicalSparseGrid due to possible complex datatypes. Returns true for DistributedSparseGrids.HierarchicalCollocationPoints which should get refined.

Arguments

rt::T : Scaling weight of DistributedSparseGrids.HierarchicalCollocationPoint. T is returntype of fun calling the OGS6 binary.

tolrt::T : Average scaling weight weighted by SparseGridParams.tol. T is returntype of fun calling the OGS6 binary.

mintol::T : Not used at the moment.

Utils userfile

OpenGeoSysUncertaintyQuantification.dependend_tensor_parameter! — Method

dependend_tensor_parameter!(modeldef::Ogs6ModelDef, stoparam::String, master_ind::Int, slave_ind::Int, depfunc)

Helper function to model a dependency of entries of a tensor parmeter such as the permeability tensor k. Uses OGS6 Pathes for parameter indentification (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath).

Arguments

modeldef::Ogs6ModelDef: OSG6 Model Definition

stoparam::String: OGS6 Path to parameter

master_ind::Int: master index

slave_ind::Int: slave index

depfunc: dependency function

Example:

dependend_tensor_parameter!(
	modeldef, 
	path_to_permeability, 
	1,
	3, 
	x->0.5*x)

The above example sets the third entry of the permeability tensor to half of the value of the first entry.

OpenGeoSysUncertaintyQuantification.dependend_parameter! — Method

dependend_parameter!(
	modeldef::Ogs6ModelDef, 
	masterstoparam::String, 
	slavestoparam::String, 
	master_ind::Int, 
	slave_ind::Int, 
	depfunc
	)

Helper function to model a dependency of two OGS6 parameter. Uses OGS6 Pathes for parameter indentification (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath).

Arguments

modeldef::Ogs6ModelDef: OSG6 Model Definition

masterstoparam::String: OGS6 Path to master parameter

slavestoparam::String: OGS6 Path to slave parameter

master_ind::Int: master index

slave_ind::Int: slave index

depfunc: dependency function

Example:

dependend_tensor_parameter!(
	modeldef, 
	path_to_permeability_id1, 
	path_to_permeability_id2, 
	1, 
	1, 
	x->x)

The above example couples two permeabilites of different material layers.