Library

Contents

Library

Contents

Project setup

Structs

General functions

Utils

Adaptive sparse grid

Sobols indices

Userfile

Typedefs

Index

Project setup

OpenGeoSysUncertaintyQuantification.jl provides methods for the generation of a stochastic OGS6 model from a validly configured OGS6 project. generatePossibleStochasticParameters scans existing OGS6 projects for possible stochastic parameters. generateStochasticOGSModell is a helper function which creates an data structure containing all parameters describing the stochastic OGS model, defined by StochasticOGSModelParams. generateSampleMethodModel generates the data structure for the sample method, defined by the abstract type SampleMethodParams. Currently implented are: SparseGridParams, MonteCarloParams, MonteCarloSobolParams, or MonteCarloMorrisParams.

OpenGeoSysUncertaintyQuantification.generatePossibleStochasticParameters — Function

generatePossibleStochasticParameters(
	projectfile::String,
	file::String="./PossibleStochasticParameters.xml",
	keywords::Vector{String}=ogs_numeric_keyvals
	)

Helper function for initial setup of a stochastic project. Scans an existing OGS6 projectfile (.prj) for keywords indicating numeric values usable as stochastic input parameter. Returns a Vector of Strings with OGS6 pathes. Uses OGS6 Pathes for parameter indentification (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath). Outputs an XML file written by XMLParser.Julia2XML.

Arguments

projectfile::String: Path to OGS6 project file (e.g. ./path/to/project.prj)

file::String: Path of output XML file which gets written by this function.

keywords::Vector{String}: List of keywords.

OpenGeoSysUncertaintyQuantification.loadStochasticParameters — Function

loadStochasticParameters(file::String="./PossibleStochasticParameters.xml")

Loads an XML file containing OGS6 pathes by XMLParser.XML2Julia.

Arguments

file::String: Path of XML file with OGS6 pathes.

OpenGeoSysUncertaintyQuantification.generateStochasticOGSModell — Function

generateStochasticOGSModell(
projectfile::String,
simcall::String,
additionalprojecfilespath::String,
postprocfile::Vector{String},
stochpathes::Vector{String},
outputpath="./Res",
stochmethod=AdaptiveHierarchicalSparseGrid,
n_local_workers=50,
userfunctionfile="./user_functions.jl",
sogsfile="StochasticOGSModelParams.xml"
)

Helper function for the generation of a stochastic OSG6 model. Writes the userfunction.jl file. Writes the stochastic model to an XML file by XMLParser.Julia2XML.

Arguments

projectfile::String: Path to OGS6 project file (e.g. ./path/to/project.prj)

simcall::String : Path to the OGS6 binary (e.g. path/to/ogs/bin/ogs).

additionalprojecfilespath::String : Path to the folder with additional project files (meshes & scripts) which gets copied to each realization folder.

postprocfiles::Vector{String} : Array of OGS6 postprocessing results containing either vtu files (readable by VTUFileHandler.VTUFile) or xdmf files (readable by XDMFFileHandler.XDMF3File).

stochpathes::Vector{String} : Vector with OGS6 pathes (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath).

outputpath::String : Output path (e.g. path/to/Res/).

stochmethod::Type : Either DistributedSparseGrids.AdaptiveHierarchicalSparseGrid, DistributedMonteCarlo.MonteCarlo, DistributedMonteCarlo.MonteCarloSobol, or DistributedMonteCarlo.MonteCarloMorris.

n_local_workers::Int : Number of local workers to be added by Distributed.addprocs.

userfunctionfile::String : path to userfunction file.

file::String : path to file to write the StochasticOGSModelParams as XML file (e.g. path/to/stochmodel.xml).

OpenGeoSysUncertaintyQuantification.generateSampleMethodModel — Function

generateSampleMethodModel(sogs::StochasticOGSModelParams, anafile="SampleMethodParams.xml")

Helper function for the creation of the SampleMethodParams data structure. Writes the output file with XMLParser.Julia2XML.

Arguments

sogs::StochasticOGSModelParams : Stochastic OGS model generated, e.g. with generateStochasticOGSModell.

anafile::String : Path to output XML.

generateSampleMethodModel(sogsfile::String, anafile="SampleMethodParams.xml")

Helper function for the creation of the SampleMethodParams data structure. Writes the output file with XMLParser.Julia2XML.

Arguments

sogs::StochasticOGSModelParams : Path to stochastic ogs xml model file (e.g. path/to/stochmodel.xml)

anafile::String : Path to output XML.

OpenGeoSysUncertaintyQuantification.writeStochasticParameters — Function

writeStochasticParameters(file::String="./PossibleStochasticParameters.xml")

Writes an XML file containing OGS6 pathes by XMLParser.XML2Julia.

Arguments

pathes::Vector{String}: Pathes to write.

file::String: Path of XML file with OGS6 pathes.

Base.write — Method

write(sogs::StochasticOGSModelParams)

Writes an XML file containing OGS6 pathes by XMLParser.XML2Julia to sogs.file.

Arguments

sogs::StochasticOGSModelParams: Stochastic OGS model.

Base.write — Method

write(file::String="./PossibleStochasticParameters.xml")

Writes an XML file containing OGS6 pathes by XMLParser.XML2Julia to sogs.file.

Arguments

sogs::SampleMethodParams: Sample method parameters

Structs

This are the main data structure needed for handling a stochastic OGS6 model. OGS6ProjectParams, StochasticOGS6Parameter, StochasticOGSModelParams, SparseGridParams, MonteCarloParams, MonteCarloSobolParams, and MonteCarloMorrisParams describing the stochastic model and containing all needed data to handle the OGS6 simulation calls. This stuctures can be loaded from and written to hard drive and manipulated with an text editor. OGSUQASG, OGSUQMC, OGSUQMCSobol, OGSUQMCMorris are the initialized stochastic models containing the data structures and run-time objects.

OpenGeoSysUncertaintyQuantification.OGS6ProjectParams — Type

mutable struct OGS6ProjectParams

Container for OpenGeoSys 6 Parameters

Fields

projectfile::String : Path to the OGS6 project file (e.g. path/to/project.prj).

simcall::String : Path to the OGS6 binary (e.g. path/to/ogs/bin/ogs).

additionalprojecfilespath::String : Path to the folder with additional project files (meshes & scripts) which gets copied to each realization folder.

outputpath::String : Path to Result folder (e.g. path/to/stochprojectfolder/Res/).

postprocfiles::Vector{String} : Array of OGS6 postprocessing results containing either vtu files (readable by VTUFileHandler.VTUFile) or xdmf files (readable by XDMFFileHandler.XDMF3File).

OpenGeoSysUncertaintyQuantification.StochasticOGS6Parameter — Type

mutable struct StochasticOGS6Parameter

Container for the definition of a stochastic OGS6 parameter. For all distributions a lower and upper bound have to be provided. For Distributions.Uniform, this can be the interval $[a,b]$, for Distributions.Normal, proper bounds have to be provided.

Fields

path::String : OGS6 path definition (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath).

valspec::Int : Value specifier (1 for scalar parameters, $i$ for $i$-th value of a tensor parameters written as a vector in OGS6 e.g. [$a_1=a_{11}, a_2=a_{12}, a_3=a_{21}, a_4=a_{22}$]).

dist::UnivariateDistribution : Univariate distribution (see Distributions.UnivariateDistribution).

lower_bound::Float64 : Lower bound for truncated distribution (see Distributions.truncated).

upper_bound::Float64 : Upper bound for truncated distribution (see Distributions.truncated).

OpenGeoSysUncertaintyQuantification.StochasticOGSModelParams — Type

mutable struct StochasticOGSModelParams

Container defining the stochastic OGS6 model.

Fields

ogsparams::OGS6ProjectParams : OGS 6 project parameters.

stochparams::Vector{StochasticOGS6Parameter} : Vector defining the stochastic state space.

samplemethod::Type : Either DistributedSparseGrids.AdaptiveHierarchicalSparseGrid, DistributedMonteCarlo.MonteCarlo, DistributedMonteCarlo.MonteCarloSobol, or DistributedMonteCarlo.MonteCarloMorris.

num_local_workers::Int : Number of local workers to be added by Distributed.addprocs.

userfunctionfile::String : path to userfunction file.

file::String : path to file to write the StochasticOGSModelParams as XML file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.SparseGridParams — Type

mutable struct SparseGridParams <: SampleMethodParams

Container for sparse grid paramters. For more information on the adaptive sparse grid, see the documentation.

Fields

N::Int : Dimension of the sparse grid, e.g. the stochastic state space.

CT::Type : Type of the collocation points, normally Float64.

RT::Type : Return type of the function fun calling the OGS6 binary. Can be customized to Float64, Vector{Float64}, Matrix{Float64}, XDMFFileHandler.XDMF3File, VTUFileHandler.VTUFile, or a custom data type which have to implement the AltInplaceOpsInterface.

pointprobs::Vector{Int} : Point set properties of the sparse grid, see DistributedSparseGrids.init and the point set example.

init_lvl::Int : Number of initial hierarchical levels before adaptive refinement starts.

maxlvl::Int : Number of additional hierarchical level after init_lvl.

tol::Float64 : tolerance for adaptive refinement.

file::String : path to file to write the StochasticOGSModelParams as XML-file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.MonteCarloParams — Type

mutable struct MonteCarloParams <: SampleMethodParams

Container for Monte Carlo parameters.

Fields

N::Int : Dimension of the sparse grid, e.g. the stochastic state space.

CT::Type : Type of the coordinates in the stochastic state space, normally Float64.

RT::Type : Return type of the function fun calling the OGS6 binary. Can be customized to Float64, Vector{Float64}, Matrix{Float64}, XDMFFileHandler.XDMF3File, VTUFileHandler.VTUFile, or a custom data type which have to implement the AltInplaceOpsInterface.

nshots::Int : Number of Monte Carlo snapshots.

tol::Float64 : tolerance (not used at the moment).

file::String : path to file to write the StochasticOGSModelParams as XML-file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.MonteCarloSobolParams — Type

mutable struct MonteCarloSobolParams <: SampleMethodParams

Container for Monte Carlo Sobol parameters.

Fields

N::Int : Dimension of the sparse grid, e.g. the stochastic state space.

CT::Type : Type of the coordinates in the stochastic state space, normally Float64.

RT::Type : Return type of the function fun calling the OGS6 binary. Can be customized to Float64, Vector{Float64}, Matrix{Float64}, XDMFFileHandler.XDMF3File, VTUFileHandler.VTUFile, or a custom data type which have to implement the AltInplaceOpsInterface.

nshots::Int : Number of Monte Carlo snapshots.

tol::Float64 : tolerance (not used at the moment).

file::String : path to file to write the StochasticOGSModelParams as XML-file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.MonteCarloMorrisParams — Type

mutable struct MonteCarloMorrisParams

Container for Monte Carlo Morris parameters.

Fields

N::Int : Dimension of the sparse grid, e.g. the stochastic state space.

CT::Type : Type of the coordinates in the stochastic state space, normally Float64.

RT::Type : Return type of the function fun calling the OGS6 binary. Can be customized to Float64, Vector{Float64}, Matrix{Float64}, XDMFFileHandler.XDMF3File, VTUFileHandler.VTUFile, or a custom data type which have to implement the AltInplaceOpsInterface.

ntrajectories::Int : Number of trajectories. Each trajectories triggers $N+1$ deterministic OGS6 runs.

lhs_sampling::Bool : If set to true a optimized Latin Hypercube design is used instead of Monte Carlo integration. Functionality provided by the LatinHypercubeSampling.jl package.

tol::Float64 : tolerance (not used at the moment).

file::String : path to file to write the StochasticOGSModelParams as XML-file by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.OGSUQParams — Type

mutable struct OGSUQParams

Container for stochastic model parameters and sample method parameters.

Fields

stochasticmodelparams::StochasticOGSModelParams : stochastic model parameters.

samplemethodparams::SampleMethodParams : sample method parameters:

Can be instantiated by

OGSUQParams(
	stochasticmodelparams::StochasticOGSModelParams, 
	samplemethodparams::SampleMethodParams
	)

or

OGSUQParams(
	file_stochasticmodelparams::String, 
	file_samplemethodparams::String
	)

where file_stochasticmodelparams and file_samplemethodparams are pathes to xml files of StochasticOGSModelParams and SampleMethodParams, respectively, written by XMLParser.Julia2XML.

OpenGeoSysUncertaintyQuantification.OGSUQASG — Type

mutable struct OGSUQASG

A stochastic OGS6 adaptive sparse grid model. The sparse grid is implemented by DistributedSparseGrids.jl. For more information, see the documentation. Instatiated by init function.

Fields

ogsuqparams::OGSUQParams : stochastic model and sample method parameters.

asg::DistributedSparseGrids.AdaptiveHierarchicalSparseGrid : instance of the adaptive sparse grid

OpenGeoSysUncertaintyQuantification.OGSUQMC — Type

mutable struct OGSUQMC

A stochastic OGS6 Monte Carlo model. The Monte Carlo Method is implemented by DistributedMonteCarlo.jl. Instatiated by init function.

Fields

ogsuqparams::OGSUQParams : stochastic model and sample method parameters.

asg::DistributedMonteCarlo.MonteCarlo : instance of the Monte Carlo integrator.

OpenGeoSysUncertaintyQuantification.OGSUQMCSobol — Type

mutable struct OGSUQMCSobol

A stochastic OGS6 Sobol model. Integration of the Sobol indices implemented by DistributedMonteCarlo.jl. Instatiated by init function.

Fields

ogsuqparams::OGSUQParams : stochastic model and sample method parameters.

mc::DistributedMonteCarlo.MonteCarloSobol : instance of the Monte Carlo Sobol integrator.

OpenGeoSysUncertaintyQuantification.OGSUQMCMorris — Type

mutable struct OGSUQMCMorris

A Stochastic OGS6 Morris model. Integration of the Sobol indices implemented by DistributedMonteCarlo.jl. Instatiated by init function.

Fields

ogsuqparams::OGSUQParams : stochastic model and sample method parameters.

mc::DistributedMonteCarlo.MonteCarloMorris : instance of the Monte Carlo Sobol integrator.

General functions

Given both necessary data structures describing a stochastic OGS6 model have been created, the following methods can be used to initialize the stochastic model and start the sampling procedure:

ogsuqparams = OGSUQParams("StochasticOGSModelParams.xml", "SampleMethodParams.xml")
ogsuqmc = OGSUQ.init(ogsuqparams)
res = OGSUQ.start!(ogsuqmc)

The data structures StochasticOGSModelParams and SampleMethodParams are automatically written to hard drive as XML files and can be manipulated in any text editor. The init function initalizes the model and adds the local workers. start! starts the sampling procedure for the adaptive sparse grid. For Monte Carlo methods only the sample points are generated upon calling [start!]. The sampling procedure, i.e. the determinitic OGS6 calls, are only executed if 𝔼 or variance is called, since for Monte Carlo methods, the results are only loaded partially into memory.

DistributedSparseGrids.init — Method

init(ogsuqparams::OGSUQParams)

Helper function to instantiate a stochastic OGS6 model. Return an object of type OGSUQASG, OGSUQMC, OGSUQMCSobol, or OGSUQMCMorris.

Arguments

ogsuqparams::OGSUQParams: Stochastic model parameters.

OpenGeoSysUncertaintyQuantification.start! — Function

start!(ogsuqasg::OGSUQASG, refinetohyperedges=false)

Starts the adaptive sparse grid sampling procedure.

Arguments

ogsuqasg::OGSUQASG: Stochastic OGS adaptive sparse grid model.

refinetohypercube::Bool: If set to true the sparse grid is refined such that all vertices of the n-cube spawn.

start!(ogsuqmc::OGSUQMC)

Starts the Monte Carlo sampling. Since Monte Carlo results are loaded each time expected value or variance is computed, only the result folder is checked for existing Monte Carlo snaphots. If there are any, the snapshot coordinates are reloaded.

Arguments

ogsuqmc::OGSUQMC: Stochastic OGS Monte Carlo model.

start!(ogsuqmc::OGSUQMCSobol)

Starts the Monte Carlo Sobol sampling.

Arguments

ogsuqmc::OGSUQMCSobol: Stochastic OGS Monte Carlo Sobol model.

start!(ogsuqmc::OGSUQMCMorris)

Starts the Monte Carlo Sobol sampling.

Arguments

ogsuqmc::OGSUQMCMorris: Stochastic OGS Monte Carlo Morris model.

OpenGeoSysUncertaintyQuantification.𝔼 — Function

𝔼(ogsuqmc::OGSUQMC)

Computes the expected value. Calls DistributedMonteCarlo.distributed_𝔼. All values are computet in-place. Only few values are stored in memory simultaneously so that the expected value of large sample sizes can be computed.

Arguments

ogsuqmc::OGSUQMC: Stochastic OGS Monte Carlo model.

𝔼(ogsuqasg::OGSUQASG)

Computes the expected value. Hereby, the physical model ogsuqasg.asg is evaluated against the multivariate pdf. Results are stored in memory multiple times. Feasible if all samples only occupy a fraction of the system memory.

Arguments

ogsuqasg::OGSUQASG: Stochastic OGS adaptive sparse grid model.

OpenGeoSysUncertaintyQuantification.variance — Function

variance(ogsuqasg::OGSUQASG)

Computes the expected value. Hereby, the physical model ogsuqasg.asg is evaluated against the multivariate pdf. Results are stored in memory multiple times. Feasible if all samples only occupy a fraction of the system memory.

Arguments

ogsuqasg::OGSUQASG: Stochastic OGS adaptive sparse grid model.

variance(ogsuqmc::OGSUQMC)

Computes the variance. Calls DistributedMonteCarlo.distributed_𝔼. All values are computet in-place. Only few values are stored in memory simultaneously so that the expected value of large sample sizes can be computed.

Arguments

ogsuqmc::OGSUQMC: Stochastic OGS Monte Carlo model.

Utils

General helper function for handling a stochastic OGS6 project.

OpenGeoSysUncertaintyQuantification.CPtoStoch — Function

CPtoStoch(x::T,stoparam::StochasticOGS6Parameter) where T<:Number

The stochastic state space is always a [-1,1]^n cube. All stochastic parameters get mapped onto this cube. Accepts an x∈[-1,1] and returns a value x'∈[stoparam.lowerbound, stoparam.upperbound].

Arguments

x::T : Value between -1 and 1.

stoparam::StochasticOGS6Parameter : stochastic ogs parameter.

OpenGeoSysUncertaintyQuantification.StochtoCP — Function

StochtoCP(x::T,stoparam::StochasticOGS6Parameter) where T<:Number

The stochastic state space is always a [-1,1]^n cube. All stochastic parameters get mapped onto this cube. Accepts an x∈[stoparam.lowerbound, stoparam.upperbound] and returns a value x'∈[-1,1].

Arguments

x::T : Value between stoparam.lowerbound and stoparam.upperbound.

stoparam::StochasticOGS6Parameter : stochastic ogs parameter.

Adaptive sparse grid

If an OGSUQASG model is used the adaptive sparse grid) serves as a surrogate model. The idea is to sample the physical state space and to integrate this state space against the multivariate pdf in an additional step. This is done because the integration of the multivariate pdf is often times much harder to integrate than the physical model, for example in the case of a multivariate normal distribution. The empirical output distributions can be integrated by empirical_cdf.

OpenGeoSysUncertaintyQuantification.scalarwise_comparefct — Method

scalarwise_comparefct(rt,tolrt,mintol)

Criterion for refinement of the DistributedSparseGrids.AdaptiveHierarchicalSparseGrid due to possible complex datatypes. Returns true for DistributedSparseGrids.HierarchicalCollocationPoints which should get refined.

Arguments

rt::T : Scaling weight of DistributedSparseGrids.HierarchicalCollocationPoint. T is returntype of fun calling the OGS6 binary.

tolrt::T : Average scaling weight weighted by SparseGridParams.tol. T is returntype of fun calling the OGS6 binary.

mintol::T : Not used at the moment.

OpenGeoSysUncertaintyQuantification.empirical_cdf — Function

empirical_cdf(
	ogsuqasg::OGSUQASG, 
	quant_vals::AbstractVector{Float64}, 
	samplepoint::Vector{Float64}, 
	MC_N::Int, 
	postprocfun::F,  
	xdmf::XDMF3File
	)

Computes the empirical cdf of an adaptive sparse grid surrogate model by Monte Carlo integration. Assumes a two-dimensional OGS6 postprocessing result as xdmf file which lies in the XY-plane. Returns the empirical expected value and the empirical output distribution.

Arguments

ogsuqasg::[OGSUQASG] : Stochastic OGS6 adaptive sparse grid model.

quant_vals::AbstractVector{Float64} : Vector with quantiles. For each value the quantile function is evalulated.

samplepoint::Vector{Float64} : Point with coordinates [x, y, 0.0].

MC_N::Int : Number of Monte Carlo snapshots for integration of the empirical output.

postprocfun::F : Postprocessing function, see sample_postproc_fun.

xdmf::XDMF3File : An arbirtrary XDMF3File from the result folder is needed with the topology of the mesh.

OpenGeoSysUncertaintyQuantification.sample_postproc_fun — Function

sample_postproc_fun(asg, x, inds, ξs, retval_proto)

A sample postprocessing function for empirical_cdf.

# copy result prototype here so that the sparse grid can interpolate without allocations
	ret = deepcopy(retval_proto) 
	# interpolate value from sparse grid
	DistributedSparseGrids.interpolate!(ret,asg,x)
	# use the fourth result value (could also be something like ret["sigma"][inds,:,:])
	# and interpolate the value at the element coordinates by shape functions
	val = Tri6_shapeFun(ξs)'*ret[4][inds]

Arguments

asg : Addaptive sparse grid

x : sample point ∈ [-1,1]^n

inds : Element indices

ξs : Element coordinates

retval_proto : Prototype for result type (only known after first OGS6 call, i.e. DistributedSparseGrids.scaling_weight(first(asg)))

Sobols indices

With OpenGeoSysUncertaintyQuantification.jl, Sobol indices of whole OGS6 postprocessing results can be integrated. To write such results to hard drive, the function write_sobol_field_result_to_XDMF and write_sobol_multifield_result_to_XDMF are used.

OpenGeoSysUncertaintyQuantification.write_sobol_field_result_to_XDMF — Function

write_sobol_field_result_to_XDMF(
	ogsuqmc::OGSUQMCSobol, 
	sobolvars, 
	fieldname::String, 
	varval, 
	expval, 
	xdmf_proto_path::String
	)

Writes the result of start!(ogsuqmcsobl::OGSUQMCSobol) to an XMDF file.

Following outputs are provided by a Monte Carlo Sobol sampling:

expval, varval, sobolvars, totsobolvars = start!(ogsuqmcsobol)

Assumes each index in sobolvars is a single field result (i.e.

xdmf["temperature_interpolated"][:,some_timestep]::Vector{Float64}

)

The result are written in the 0-th time slice of the XDMF3File.

Arguments

ogsuqmc::OGSUQMCSobol : Instance of OGSUQMCSobol.

sobolvars : A result of start!(ogsuqmcsobl::OGSUQMCSobol).

fieldname::String : Fieldname, e.g. "temperature".

varval : Global variance of stochastic state space defined in ogsuqmc.

expval : Exptected value of stochastic state space defined in ogsuqmc.

xdmf_proto_path::String : Path to a xdmf file with similar topology like the result field.

OpenGeoSysUncertaintyQuantification.write_sobol_multifield_result_to_XDMF — Function

write_sobol_multifield_result_to_XDMF(
	ogsuqmc::OGSUQMCSobol, 
	sobolvars,
	field, 
	fieldname::String, 
	varval, 
	expval, 
	xdmf_proto_path::String
	)

Writes the result of start!(ogsuqmcsobl::OGSUQMCSobol) to an XMDF file.

Following outputs are provided by a Monte Carlo Sobol sampling:

expval, varval, sobolvars, totsobolvars = start!(ogsuqmcsobol)

Assumes each index in sobolvars is a multifield result (i.e.

[xdmf["temperature_interpolated"][:,some_timestep]::Vector{Float64}, xdmf["pressure_interpolated"][:,some_timestep]::Vector{Float64}]

)

The result are written in the 0-th time slice of the XDMF3File.

Arguments

ogsuqmc::OGSUQMCSobol : Instance of OGSUQMCSobol.

sobolvars : A result of start!(ogsuqmcsobl::OGSUQMCSobol).

field::Any : Field index to be postprocessed. Has to be compatible with getindex, i.e. map(x->x[field] sobolvars).

fieldname::String : Fieldname, e.g. "temperature".

varval : Global variance of stochastic state space defined in ogsuqmc.

xdmf_proto_path::String : Path to a xdmf file with similar topology like the result field.

Userfile

Upon the generation of the stochastic model, a user file (userfile.jl) is generated as well. This userfile executes the OGS6 simulation call. The stochastic model as well as the quantities of interest can be altered in this file. In case of a altered quantity of interest, the return type in [SampleMethodParams] has to be altered, accordingly.

OpenGeoSysUncertaintyQuantification.dependend_tensor_parameter! — Method

dependend_tensor_parameter!(
	modeldef::Ogs6ModelDef, 
	stoparam::String, 
	master_ind::Int, 
	slave_ind::Int, 
	depfunc
)

Helper function to model a dependency of entries of a tensor parmeter such as the permeability tensor k. Uses OGS6 Pathes for parameter indentification (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath).

Arguments

modeldef::Ogs6ModelDef: OSG6 Model Definition

stoparam::String: OGS6 Path to parameter

master_ind::Int: master index

slave_ind::Int: slave index

depfunc: dependency function

Example:

dependend_tensor_parameter!(
	modeldef, 
	path_to_permeability, 
	1,
	3, 
	x->0.5*x
)

The above example sets the third entry of the permeability tensor to half of the value of the first entry.

OpenGeoSysUncertaintyQuantification.dependend_parameter! — Method

dependend_parameter!(
	modeldef::Ogs6ModelDef, 
	masterstoparam::String, 
	slavestoparam::String, 
	master_ind::Int, 
	slave_ind::Int, 
	depfunc
)

Helper function to model a dependency of two OGS6 parameter. Uses OGS6 Pathes for parameter indentification (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath).

Arguments

modeldef::Ogs6ModelDef: OSG6 Model Definition

masterstoparam::String: OGS6 Path to master parameter

slavestoparam::String: OGS6 Path to slave parameter

master_ind::Int: master index

slave_ind::Int: slave index

depfunc: dependency function

Example:

dependend_tensor_parameter!(
	modeldef, 
	path_to_permeability_id1, 
	path_to_permeability_id2, 
	1, 
	1, 
	x->x
)

The above example couples two permeabilites of different material layers.

OpenGeoSysUncertaintyQuantification.setStochasticParameter! — Function

setStochasticParameter!(
	modeldef::Ogs6ModelDef, 
	stoparam::StochasticOGS6Parameter, 
	x, 
	user_func::Function,
	cptostoch::Function=CPtoStoch
)

Replaces a stochastic parameter at x in the modeldef. Applies the user_func for addiational transformation (e.g. x->exp(x) in case of a lognormal distribution).

Arguments

modeldef::[Ogs6ModelDef] : OGS6 model definition.

stoparam::StochasticOGS6Parameter : stochastic ogs parameter.

x::Float64 : sample point between -1 and 1. Gets transformed by cptostoch.

user_func::Function : User function for additional transformation.

cptostoch::Function : Transformation from [-1,1] to [stoparam.lower_bound, stoparam.upper_bound].

OpenGeoSysUncertaintyQuantification.setStochasticParameters! — Function

setStochasticParameters!(
	modeldef::Ogs6ModelDef, 
	stoparams::Vector{StochasticOGS6Parameter}, 
	x, 
	user_funcs::Vector{Function},
	cptostoch::Function=CPtoStoch
)

Replaces all stochastic parameter at x in the modeldef. Applies all user_func for addiational transformation (e.g. x->exp(x) in case of a lognormal distribution).

Arguments

modeldef::[Ogs6ModelDef] : OGS6 model definition.

stoparam::StochasticOGS6Parameter : stochastic ogs parameter.

x::Vector{Float64} : sample point in the stochastic state space. Gets transformed by cptostoch.

user_funcs::Vector{Function} : User function for additional transformation.

cptostoch::Function : Transformation from [-1,1] to [stoparam.lower_bound, stoparam.upper_bound].

Typedefs

Here are some supertypes used for dispatching.

OpenGeoSysUncertaintyQuantification.SampleMethodParams — Type

abstract type SampleMethodParams

Supertype for Sample method parameters. Implemented by, e.g., SparseGridParams or MonteCarloParams.

OpenGeoSysUncertaintyQuantification.AbstractOGSUQ — Type

abstract type AbstractOGSUQ

Supertype for stochastic OGS6 models. Implemented by OGSUQASG.

OpenGeoSysUncertaintyQuantification.AbstractOGSUQMonteCarlo — Type

abstract type AbstractOGSUQMonteCarlo

Supertype for stochastic OGS6 models. Implemented by OGSUQMC.

OpenGeoSysUncertaintyQuantification.AbstractOGSUQSensitivity — Type

abstract type AbstractOGSUQSensitivity

Supertype for stochastic OGS6 models. Implemented by OGSUQMCSobol and OGSUQMCMorris.

Index

OpenGeoSysUncertaintyQuantification.AbstractOGSUQ

OpenGeoSysUncertaintyQuantification.AbstractOGSUQMonteCarlo

OpenGeoSysUncertaintyQuantification.AbstractOGSUQSensitivity

OpenGeoSysUncertaintyQuantification.MonteCarloMorrisParams

OpenGeoSysUncertaintyQuantification.MonteCarloParams

OpenGeoSysUncertaintyQuantification.MonteCarloSobolParams

OpenGeoSysUncertaintyQuantification.OGS6ProjectParams

OpenGeoSysUncertaintyQuantification.OGSUQASG

OpenGeoSysUncertaintyQuantification.OGSUQMC

OpenGeoSysUncertaintyQuantification.OGSUQMCMorris

OpenGeoSysUncertaintyQuantification.OGSUQMCSobol

OpenGeoSysUncertaintyQuantification.OGSUQParams

OpenGeoSysUncertaintyQuantification.SampleMethodParams

OpenGeoSysUncertaintyQuantification.SparseGridParams

OpenGeoSysUncertaintyQuantification.StochasticOGS6Parameter

OpenGeoSysUncertaintyQuantification.StochasticOGSModelParams

Base.write

Base.write

DistributedSparseGrids.init

OpenGeoSysUncertaintyQuantification.CPtoStoch

OpenGeoSysUncertaintyQuantification.StochtoCP

OpenGeoSysUncertaintyQuantification.dependend_parameter!

OpenGeoSysUncertaintyQuantification.dependend_tensor_parameter!

OpenGeoSysUncertaintyQuantification.empirical_cdf

OpenGeoSysUncertaintyQuantification.generatePossibleStochasticParameters

OpenGeoSysUncertaintyQuantification.generateSampleMethodModel

OpenGeoSysUncertaintyQuantification.generateStochasticOGSModell

OpenGeoSysUncertaintyQuantification.loadStochasticParameters

OpenGeoSysUncertaintyQuantification.sample_postproc_fun

OpenGeoSysUncertaintyQuantification.scalarwise_comparefct

OpenGeoSysUncertaintyQuantification.setStochasticParameter!

OpenGeoSysUncertaintyQuantification.setStochasticParameters!

OpenGeoSysUncertaintyQuantification.start!

OpenGeoSysUncertaintyQuantification.variance

OpenGeoSysUncertaintyQuantification.writeStochasticParameters

OpenGeoSysUncertaintyQuantification.write_sobol_field_result_to_XDMF

OpenGeoSysUncertaintyQuantification.write_sobol_multifield_result_to_XDMF

OpenGeoSysUncertaintyQuantification.𝔼